Initial commit of control console plus install tweaks
- local and remote check scripts to verify installation
- status poller and dashboard
- mass file copy, file update, and execution scripts
- timezone and locale fixes

+++ os/scripts/kernel_upgrade.sh
@@ -1,10 +1,17 @@
-#!/bin/sh
+#!/bin/bash
 
-VER=v3.8.13-bone22
+VER=3.8.13-bone22
 
-rm -f install-me.sh
-wget http://rcn-ee.net/deb/wheezy-armhf/${VER}/install-me.sh
-chmod 755 install-me.sh
-./install-me.sh
-rm -f install-me.sh
+if [ $VER == `uname -r` ] ; then
+  echo Kernel already upgraded
+else
+  rm -f install-me.sh
+  wget http://rcn-ee.net/deb/wheezy-armhf/v${VER}/install-me.sh
+  chmod 755 install-me.sh
+  ./install-me.sh
+  rm -f install-me.sh
+  if [ $1 == `-r` ] ; then
+    reboot
+  fi
+fi
